+++ Magic/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 # 7.10.6
 
+ - Fix errors using `/mgive someitem someamount` without a player target
+
 # 7.10.5
 
  - Wands are once again unstashable by default. Set `unstashable: false` on base_bound to turn this back off